fix softmax

+++ src/helper.py
@@ -20,4 +20,5 @@ def softmax(net_i, arr_net):
     net_sum = 0
     for i in arr_net:
         net_sum += exp(i)
-    return (exp(net_i))/net_sum
+
+    return float(exp(net_i)/net_sum)
